requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.7
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.5
django-braces==1.14.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==20.1.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.1
numpy==1.19.2
openpyxl==3.0.7
pandas==1.1.3
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
xlrd==1.2.0
XlsxWriter==1.3.6
